refactor: changed hover affect on the carousel cards
