+++ client/src/components/PairingSelection/PairingSelection.tsx
@@ -128,7 +128,7 @@ function PairingSelection({ onSelect, availableOptions }: PairingSelectionProps)
                 _hover={{
                   borderColor: 'brand.secondary',
                   color: 'brand.secondary',
-                  animation: `${shake} 0.4s ease-in-out`,
+                  animation: `pulseGlow 2s ease-in-out`,
                 }}
               >
                 {option}
